moving function out of views
